move CellToggle to the components folder
