+++ source/views/components/filter/section-toggle.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import type {ToggleType} from './types'
 import {Section} from 'react-native-tableview-simple'
-import {CellToggle} from '../../components/cell-toggle'
+import {CellToggle} from '../../components/cells/toggle'
 
 type PropsType = {
   filter: ToggleType,
